requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# commands.py: 3
Jinja2 == 2.9.6

# ANFPWebScraper.py: 3
beautifulsoup4 == 4.6.0

# settings.py: 2
python_dotenv == 0.6.5

# futbot.py: 4
python_telegram_bot == 7.0.1

# ANFPWebScraper.py: 2
requests == 2.18.1
